Fixed user user can delete review in their account, fixed account page

+++ core/views.py
@@ -158,12 +158,10 @@ def business_detail(request, slug):
                 review.reviewer = request.user
                 review.save()
                 return redirect('business_detail', slug=business_review.slug)
-    
         else:
             messages.success(request, "Oops , must be logged in to review! ")
 
     review = LeaveReview.objects.filter(business_review=business_review)
-    # average_score = review.aggregate(Avg('rating'))
 
     return render(
         request, 'bsbusiness_detail.html', {
@@ -205,10 +203,10 @@ def newbusiness_detail(request, slug):
 @login_required
 def user_delete_review(request, id):
     user = request.user
-    review = LeaveReview.objects.filter(reviewer=user)
-
-    review.delete()
-    return redirect('home')
+    review = LeaveReview.objects.get(id=id)
+    if (request.user == review.reviewer):
+        review.delete()
+    return redirect('get_user_profile')
 
 
 @login_required
@@ -216,11 +214,13 @@ def get_user_profile(request):
     user = request.user
 
     reviews = LeaveReview.objects.filter(reviewer=user)
-    favorite_event = user.favorite.all()
+    favorite_event = user.favorite.all
+   
 
     return render(request, 'bsuser_account.html', {
         'reviews': reviews,
         'favorite_event': favorite_event,
+        
     })
